requirements.txt
Jinja2==3.0.1
Flask==2.0.1
selenium==3.141.0
sib_api_v3_sdk==7.2.0
msedge_selenium_tools==3.141.3
google_api_python_client==2.21.0
protobuf==3.18.0
google-cloud-firestore
gunicorn==20.1.0